requirements.txt
appdirs==1.4.2
cffi==1.9.1
click==6.7
cryptography==1.7.2
Flask==0.12
Flask-Autodoc==0.1.2
Flask-MongoAlchemy==0.7.2
Flask-MongoKit==0.6
gunicorn==19.6.0
idna==2.4
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
MongoAlchemy==0.21
mongokit==0.9.1.1
packaging==16.8
pyasn1==0.2.3
pycparser==2.17
PyJWT==1.4.2
pymongo==3.4.0
pyparsing==2.1.10
six==1.10.0
Werkzeug==0.11.15
